Move statistics to a new usage page
