+++ interface/app/$libraryId/settings/client/index.ts
@@ -3,6 +3,7 @@ import { RouteObject } from 'react-router';
 export default [
 	{ path: 'general', lazy: () => import('./general') },
 	{ path: 'account', lazy: () => import('./account') },
+	{ path: 'usage', lazy: () => import('./usage')},
 	{ path: 'appearance', lazy: () => import('./appearance') },
 	{ path: 'keybindings', lazy: () => import('./keybindings') },
 	{ path: 'extensions', lazy: () => import('./extensions') },
