+++ interface/components/AuthRequiredOverlay.tsx
@@ -7,7 +7,7 @@ export function AuthRequiredOverlay() {
 
 	if (authState.status !== 'loggedIn')
 		return (
-			<div className="absolute inset-0 z-50 flex items-center justify-center bg-app/75 backdrop-blur-sm">
+			<div className="absolute inset-0 z-50 flex items-center justify-center rounded-md bg-app/75 backdrop-blur-sm">
 				{authState.status === 'loading' || authState.status === 'loggingIn' ? (
 					<Loader />
 				) : (
